refactor: rename gh-auth module
